+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.7
+
+- [windows] Replace `win32` with `win32_registry`
+
+## 0.1.6
+
+- [windows] fix: use correct max length of registry key #
+
 ## 0.1.5
 
 - [linux] Fixed wrong app name when writing .desktop file. #4